Add enumerate_* and indices_* methods
It can often be often be convenient to have not only the elements of
an array but also the indices. For a `Vec`, this is simple--simply
call `vec.iter().enumerate()`. However, this does not work with the
two-dimensional indices of an `Array2D`. The `enumerate_*` methods
return an iterator over tuples of index and value.

The `indices_*` methods provide iterators over just the indices.