Add map methods
Previously, the only way to create a new `Array2D` was to either
manually specify both the source and size of the array or to provide
preexisting rows/columns. However, it is common to want to create a
new array from the contents of a previously existing array.

The new `map_*` methods allow the user to provide a "mapping" function
that transforms the elements of the existing array into new elements
for the new array.

There are "row major" and "column major" versions of these
methods. The differences will not be relevant unless the provided
"mapping" function is modifying shared state. See the test cases that
modify a "counter" variable (`mutable_counter`).